Main Functions

Note

All functions support as input Unitful.Quantity values for frequency (e.g. Hz), angles (° or rad) and length (e.g. m, but only for values which are expected in km like the location altitude or the rain height hᵣ)

The primary interface for calculating atmospheric attenuations.

Primary Function

ITUPropagationModels.ItuRP618.attenuations — Function

attenuations(latlon, f, el, p; D, η, alt, polarization, polarization_angle, warn, extra_kwargs...)
attenuations(lat::Number, lon::Number, args...; kwargs...)

Computes the various attenuations for earth/space links based on Section 2.4.1 of ITU-R P618-13.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

f: frequency (GHz)

el: elevation angle (degrees)

p: exceedance probability (%)

Keyword Arguments

D: antenna diameter (m)

η (also accepted as efficiency): antenna efficiency (% from 0 to 100, defaults to 60).

alt: altitude of ground station (km). Defaults to ItuRP1511.topographicheight(latlon)

polarization: polarization (EnumHorizontalPolarization, EnumVerticalPolarization, or EnumCircularPolarization)

polarization_angle: tilt angle [degrees] of the electric field polarization w.r.t. horizontal polarization. Note: This field is computed from the polarization argument if not provided, but this function disregards the polarization argument if polarization_angle is explicitly provided.

warn: Whether to warn if the inputs are outside the supported range. Defaults to !SUPPRESS_WARNINGS[]

See extended help for more details on extra keyword arguments and for maximizing the speed of the computation when some intermediate inputs are known.

Return

(Ac=Ac, Ag=Ag, Ar=Ar, As=As, At=At): cloud, gas, rain, scintillation, total attenuations (dB)

Extended help

This function accepts some additional keyword arguments that are intermediate inputs to the computation, and can be provided to maximize the speed of the computation. These extra_kwargs are:

Ag_zenith: Zenith gaseous attenuation (dB/km) at zenith (90° elevation)

Ac_zenith: Zenith cloud attenuation (dB/km) at zenith (90° elevation)

Nwet: Wet term surface refractivity (N-units)

R001: Annual rain rate [mm/h] exceeded 0.01% of the time.

hᵣ: Rain height [Km] to be used for the computation.

gamma_oxygen (also accepted as γₒ): Specific attenuation due to oxygen (dB/km) computed from the average surface conditions, at the desired location. Note: This is used to speed up the computation of Ag_zenith but is ignored if the Ag_zenith argument is provided

Additionally, the function can be called with Val(true) as last positional argument (after p) to also return the various intermediate inputs used in the computation for speeding up subsequent calls where only the elevation angle changes. Here is an example on how to use this when needing to compute attenuations for a given ground location towards multiple satellites at different elevation angles:

Fast computation example

els = [10, 20, 30, 40, 50, 60, 70, 80, 90]
ll = LatLon(0,0)
p = .5 # .5% outage
D = 1 # 1m antenna diameter
# We compute the attenuations for the first elevation angle and save also the various intermediate values stored in the `kwargs` field of the output.
nt = ItuRP618.attenuations(ll, 30, first(els), p, Val(true); D)
# We create a vector of outputs for different elevations exploiting the precomputed kwargs
outs = map(eachindex(els)) do i
    i == 1 && return nt.attenuations
    ItuRP618.attenuations(ll, 30, els[i], p; nt.kwargs...)
end

ITU-R P.453 - Refractivity

ITUPropagationModels.ItuRP453.wettermsurfacerefractivityannual — Function

wettermsurfacerefractivityannual(latlon, p)
wettermsurfacerefractivityannual(lat::Number, lon::Number, args...; kwargs...)

Interpolates wet term of the surface refractivity at an exceedance probability p provided as input in % based on Section 2.2.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p: exceedance probability (%)

Return

Nwet::Real: wet term of the surface refractivity (ppm)

ITUPropagationModels.ItuRP453.wettermsurfacerefractivityannual_50 — Function

wettermsurfacerefractivityannual_50(latlon)
wettermsurfacerefractivityannual_50(lat::Number, lon::Number, args...; kwargs...)

Returns the wet term of surface refractivity at 50% exceedance probability (Based on annual data) at the desired location

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

ITU-R P.618 - Rain Attenuation and Scintillation

ITUPropagationModels.ItuRP618.rainattenuation — Function

rainattenuation(latlon, f, el, p; polarization::IturEnum=EnumCircularPolarization, kwargs...)
rainattenuation(lat::Number, lon::Number, args...; kwargs...)

Computes rain attenuation based on Section 2.2.1.1.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

f::Real: frequency (GHz),

el::Real: elevation angle (degrees)

p::Real: exceedance probability (%)

Keyword Arguments

h_r or hᵣ: Rain height [Km] to be used for the computation. Defaults to ItuRP1511.topographicheight(lat, lon)

polarization::IturEnum=EnumCircularPolarization: polarization (EnumHorizontalPolarization, EnumVerticalPolarization, or EnumCircularPolarization)

Note that this last argument is overridden by the keyword argument polarization_angle if provided

polarization_angle: Tilt angle [degrees] of the electric field polarization w.r.t. horizontal polarization. Defaults to 45, corresponding to a circularly polarized field.

alt: Altitude [km] of the receiver. Defaults to ItuRP1511.topographicheight(latlon)

R001: Annual rain rate [mm/h] exceeded 0.01% of the time. Defaults to ItuRP837.rainfallrate001(latlon)

warn: Whether to warn if the inputs are outside the supported range. Defaults to !SUPPRESS_WARNINGS[]

Return

Ap::Float64: rain attenuation (dB)

ITUPropagationModels.ItuRP618.scintillationattenuation — Function

scintillationattenuation(latlon::LatLon, f::Real, el::Real, p::Real; D::Real=1.0, η::Real=60.0)
scintillationattenuation(lat::Number, lon::Number, args...; kwargs...)

Computes scintillation attenuation based on Section 2.4.1.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

f::Real: frequency (GHz)

el::Real: elevation angle (degrees)

p::Real: exceedance probability (%)

Keyword Arguments

D::Real=1.0: antenna diameter (m)

η::Real=60: antenna efficiency (% from 0 to 100, typically 60). Can also be provided with the name efficiency.

warn: Whether to warn if the inputs are outside the supported range. Defaults to !SUPPRESS_WARNINGS[]

Return

Ascintillation::Real: scintillation attenuation (dB)

ITU-R P.676 - Gaseous Attenuation

ITUPropagationModels.ItuRP676.gaseousattenuation — Function

Ag = gaseousattenuation(latlon, f, el, p; alt = nothing, gamma_oxygen = nothing, γₒ = nothing)
gaseousattenuation(lat::Number, lon::Number, args...; kwargs...)

Computes the statistical gaseous attenuation for a slant path following the approximate computation specified in Annex 2 of ITU-R P.676-13.

More specifically this computes $Ag = Ao + Aw$ implementing:

Equation 32 in Section 1.2 for Oxygen attenuation Ao

Equation 40 in Section 2.3 for Water vapour attenuation Aw

Arguments

latlon: Object specifying the latitude and longitude of the location of interest, must be an object that can be converted to an instance of ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

f: frequency (GHz)

el: elevation angle (degrees)

Keyword arguments

alt: Altitude at the provided location, to be used for computing the various intermediate varaibles. If not provided, default to the altitude computed with ItuRP1511.topographicheight

gamma_oxygen (or γₒ): Specific attenuation due to oxygen (dB/km) computed from the average surface conditions, at the desired location. This is computed automatically based on other inputs if not explicitly provided, but it is only location dependent and >70% of the time is spent in computing this, so consider precomputing and passing it directly for maximum speed

See the extended help for the signature of the function with precomputed intermediate variables.

Extended help

Alternative method

When the values (both average and ccdf) of Pressure, Temeprature and water vapour density are already available, the computation can be made faster by using this alternative method that takes all the useful parameters as keyword arguments.

This method can also be used to simulated oxygen/gas attenuations based on instantaneous values of the surface parameters as per Sections 1.1 and 2.2 of Annex 2. In this case, the instantaneous values of P, T and ρ should be explicitly provided for the ccdf values with the same name, but also to the kwargs related to mean values (i.e. P̄, T̄ and ρ̄ respectively).

gaseousattenuation(latlon, f, el; kwargs...)

An additional method is available, for cases where the statistical values of the surface parameters are already available and do not need to be computed internally.

This method do not accept the outage probability p as last argument but expects he following keyword arguments specifying the various surface paramters (both in mean value and in ccdf statistical value):

P_mean (or P̄): Average surface total pressure (hPa) at the desired location. In case this method is used for simulating attenuation based on instantaneous values, the instantaneous total pressure P should be provided also as value to this kwarg.

T_mean (or T̄): Average surface temperature (K) at the desired location. In case this method is used for simulating attenuation based on instantaneous values, the instantaneous temperature T should be provided also as value to this kwarg.

rho_mean (or ρ̄): Average surface water vapour density (g/m^3) at the desired location. In case this method is used for simulating attenuation based on instantaneous values, the instantaneous water vapour density ρ should be provided also as value to this kwarg.

P: Surface total pressure (hPa) at the desired exceedance probability, at the desired location.

T: Surface temperature (K) at the desired exceedance probability, at the desired location.

rho: Surface water vapour density (g/m^3) at the desired exceedance probability, at the desired location.

V: Surface water vapour content (kg/m^2) at the desired exceedance probability, at the desired location.

gamma_oxygen (or γₒ): Specific attenuation due to oxygen (dB/km) computed from the average surface conditions, at the desired location. This is computed automatically based on other inputs if not explicitly provided, but it is only location dependent and >70% of the time is spent in computing this, so consider precomputing and passing it directly for maximum speed

ITU-R P.835 - Standard Atmospheres

ITUPropagationModels.ItuRP835.standardtemperature — Function

standardtemperature(Z::Float64)

Standard temperature for geometric heights <= 100 km, based on equations 2a-2g and 4a-4b of Section 1.1

Arguments

Z::Float64: geometric height (km)

Return

temperature (°K)

ITUPropagationModels.ItuRP835.standardpressure — Function

standardpressure(Z)

Standard pressure for geometric heights <= 100 km, based on equations 3a-3g and 5 of Section 1.1.

Arguments

Z: geometric height (km)

Return

dry pressure (hPa)

ITUPropagationModels.ItuRP835.standardwatervapourdensity — Function

standardwatervapourdensity(Z; kwargs...)

Standard water vapour density for geometric heights <= 100 km, based on equations 6-8 of Section 1.2.

Arguments

Z: geometric height (km)

Keyword arguments

rho_0::Real=7.5: standard ground level water vapour density (g/m^3). Can also be provided as ρ₀.

T::Real: Temperature (°K) at given height. Computed with standardtemperature(Z) if not provided.

P::Real: Total barometric pressure (hPa) at given height. Computed with standardpressure(Z) if not provided.

ITU-R P.837 - Rainfall Rate

ITUPropagationModels.ItuRP837.rainfallrate001 — Function

rainfallrate001(latlon)
rainfallrate001(lat::Number, lon::Number)

Computes rainfall rate exceeded 0.01% via bi-linear interpolation as described in Annex 1.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

Return

R::Float64: annual rainfall rate exceeded 0.01%

ITU-R P.838 - Rain Specific Attenuation

ITUPropagationModels.ItuRP838.rainspecificattenuation — Function

rainspecificattenuation(f, el; R, polarization::IturEnum, polarization_angle = nothing)

Computes rain specific attenuation for horizontal polarization based on equation 1 of Section 1.

Arguments

f::Float64: frequency (GHz)

el:Float64: path elevation angle (degrees)

Keyword Arguments

R: rain rate (mm/hr)

polarization::IturEnum: polarization (EnumHorizontalPolarization, EnumVerticalPolarization, or EnumCircularPolarization). Defaults to circular polarization.

polarization_angle: Polarization tilt angle (degrees) relative to horizontal polarization (τ = 45° for circular polarization)

Note: If this is provided, the polarization argument is ignored

Return

specific attenuation at given rain rate (dB)

ITU-R P.839 - Rain Height

ITUPropagationModels.ItuRP839.rainheightannual — Function

rainheightannual(latlon::LatLon)
rainheightannual(lat::Number, lon::Number)

Computes rain height based on the equation in Section 2. h0 will be interpolated for the given latitude and longitude.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

Return

hR: mean annual rain height (km) above mean sea level

ITUPropagationModels.ItuRP839.isothermheight — Function

isothermheight(latlon)
isothermheight(lat::Number, lon::Number)

Calculates isothermic height based on bilinear interpolation. h0 will be interpolated for the given latitude and longitude.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

Return

h0: mean annual 0°C isotherm height (km) above mean sea level

ITU-R P.840 - Cloud and Fog Attenuation

ITUPropagationModels.ItuRP840.cloudattenuation — Function

cloudattenuation(latlon, f, elevation, p)
cloudattenuation(lat::Number, lon::Number, args...; kwargs...)

Computes annual cloud attenuation along a slant path based on Section 3.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

f: frequency (GHz)

el: elevation angle (degrees)

p: exceedance probability (%)

Return

Acloud::Real: slant path cloud attenuation (dB)

ITUPropagationModels.ItuRP840.liquidwatercontent — Function

liquidwatercontent(latlon, p)
liquidwatercontent(lat::Number, lon::Number, args...; kwargs...)

Computes the integrated liquid water content at a given location and exceedance probability based on the digital annual maps in Part 1 of the Recommendation P.840-8.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p: exceedance probability (%)

ITU-R P.1511 - Topographic Data

ITUPropagationModels.ItuRP1511.topographicheight — Function

topographicheight(latlon)
topographicheight(lat::Number, lon::Number)

Calculates topographic height as per Section 1.1 of ITU-R P.1511-3.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

Return

I::Real: height (km)

ITU-R P.2145 - Surface Meteorological Data

ITUPropagationModels.ItuRP2145.surfacetemperatureannual — Function

T̄ₛ = surfacetemperatureannual(latlon; alt = nothing)
Tₛ(p) = surfacetemperatureannual(latlon, p::Real; alt = nothing)
surfacetemperatureannual(lat::Number, lon::Number, args...; kwargs...)

Computes annual surface temperature based Section 2 of the P2145-0 Recommendation and assuming the surface to be located at alt km above sea level.

If the function is called with the LatLon target position as sole positional argument, the function will return the mean surface temperature at the target location following the procedure described in Section 2.2 of the P2145-0 Recommendation. If the optional second argument p is provided, this is interpreted as the target exceedance probability and the function will follow the procedure described in Section 2.1 of the P2145-0 Recommendation.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p::Real: exceedance probability (%)

Keyword Arguments

alt::Union{Nothing, Real}: altitude (km). If provided as nothing (default) this will be computed based on the location and following Recommendation P1511-3

Returns

T̄ₛ::Float64 or Tₛ(p)::Float64: computed annual surface temperature (°K)

ITUPropagationModels.ItuRP2145.surfacewatervapourdensityannual — Function

ρ̄ₛ = surfacewatervapourdensityannual(latlon; alt = nothing)
ρₛ(p) = surfacewatervapourdensityannual(latlon, p::Real; alt = nothing)
surfacewatervapourdensityannual(lat::Number, lon::Number, args...; kwargs...)

Computes annual surface water vapour density based Section 2 of the P2145-0 Recommendation and assuming the surface to be located at alt km above sea level.

If the function is called with the LatLon target position as sole positional argument, the function will return the mean surface water vapour density at the target location following the procedure described in Section 2.2 of the P2145-0 Recommendation. If the optional second argument p is provided, this is interpreted as the target exceedance probability and the function will follow the procedure described in Section 2.1 of the P2145-0 Recommendation.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p::Real: exceedance probability (%)

Keyword Arguments

alt::Union{Nothing, Real}: altitude (km). If provided as nothing (default) this will be computed based on the location and following Recommendation P1511-3

Returns

ρ̄ₛ::Float64 or ρₛ(p)::Float64: computed annual surface water vapour density (g/m^3)

ITUPropagationModels.ItuRP2145.surfacepressureannual — Function

P̄ₛ = surfacepressureannual(latlon; alt = nothing)
Pₛ(p) = surfacepressureannual(latlon, p::Real; alt = nothing)
surfacepressureannual(lat::Number, lon::Number, args...; kwargs...)

Computes annual surface total barometric pressure based Section 2 of the P2145-0 Recommendation and assuming the surface to be located at alt km above sea level.

If the function is called with the LatLon target position as sole positional argument, the function will return the mean total barometric pressure at the target location following the procedure described in Section 2.2 of the P2145-0 Recommendation. If the optional second argument p is provided, this is interpreted as the target exceedance probability and the function will follow the procedure described in Section 2.1 of the P2145-0 Recommendation.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p::Real: exceedance probability (%)

Keyword Arguments

alt::Union{Nothing, Real}: altitude (km). If provided as nothing (default) this will be computed based on the location and following Recommendation P1511-3

Returns

P̄ₛ::Float64 or Pₛ(p)::Float64: computed annual total barometric pressure (hPa)

ITUPropagationModels.ItuRP2145.surfacewatervapourcontentannual — Function

V̄ₛ = surfacewatervapourcontentannual(latlon; alt = nothing)
Vₛ(p) = surfacewatervapourcontentannual(latlon, p::Real; alt = nothing)
surfacewatervapourcontentannual(lat::Number, lon::Number, args...; kwargs...)

Computes annual surface integrated water vapour content based Section 2 of the P2145-0 Recommendation and assuming the surface to be located at alt km above sea level.

If the function is called with the LatLon target position as sole positional argument, the function will return the mean integrated water vapour content at the target location following the procedure described in Section 2.2 of the P2145-0 Recommendation. If the optional second argument p is provided, this is interpreted as the target exceedance probability and the function will follow the procedure described in Section 2.1 of the P2145-0 Recommendation.

Arguments

latlon: object representing latitude and longitude, must be convertible to ITUPropagationModels.LatLon

This function can also be called with separate latitude and longitude as first two arguments lat and lon as per last method in the signatures above.

p::Real: exceedance probability (%)

Keyword Arguments

alt::Union{Nothing, Real}: altitude (km). If provided as nothing (default) this will be computed based on the location and following Recommendation P1511-3

Returns

V̄ₛ::Float64 or Vₛ(p)::Float64: computed annual integrated water vapour content (g/m^2)

ITUPropagationModels.ItuRP2145.annual_surface_values — Function

annual_surface_values(latlon[, p]; alt = nothing)

This function is used to provide the annual surface values for variables refrenced in P676 and P618:

P: The total barometric pressure

T: The surface temperature

ρ: The surface water vapour density

The function can be called with the outage probability p as second positional argument to compute the statsitical values, or without to compute the mean values.

It also compute the altitude of the provided location if provided as nothing as kwarg